requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mysql-connector-python==9.3.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
